scraper/requirements.txt
# Scrapy framework
Scrapy==2.11.2

# Database interaction
supabase==2.5.1
httpx==0.27.0

# Environment variables
python-dotenv==1.0.1
